rename synapse file to synapse database
